Add layoutmenu patch
It is helpful to add the `layoutmenu.sh` to a common directory where all
dwm related scripts can live. It is also required to add this common
directory to the PATH BEFORE starting dwm. So `/etc/lightdm/Xsesssion`
is a good place to do so. `Xsession` is also a nice place to restore your
wallpaper using `feh`